+++ deltafi-ui/CHANGELOG.md
@@ -3,6 +3,11 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [Unreleased]
+
+### Changed
+- Added Pretty Print to content viewer for JSON and XML (#84)
+
 ## [0.95.2]
 
 ### Changed